Allow hash or remote HTTPS url to be specified for Handler cert
This allows for the meterpreter HTTPS verification payloads to be
generated without having access to the PEM that the server is using. The
idea here is that people can generate Meterpreter payloads with SSL
certifice verification enabled without having to have direct access to
the PEM on the file system (which would require elevated access). This
means that users can specify the fingerprint manually. They can also
point the payload generator at an HTTPS server that's fronting MSF (such
as an NGINX intance) and use the SSL fingerprint from that certificate
instead. This means that hiding behind NGINX is supported with cert
validation.