Modify dll generation to make use of PE modification
This allows for DLLs to be modified in the same way that EXE files are,
allowing us to add sections, etc. This means that we can now add
payloads that are much bigger than the typical "sub" replacement method,
and so with this change we're able to generate DLLs with stageless
payloads in them.

Fixes #11170